Add HTTP status to error returned from `Do()`
The issue prompting this change was the need to inform calling services that the error was transient. Our particular case was Facebook throwing HTTP 503 in return to our calls. While the other Facebook error data is returned up the chain, we lose the context that the error happened but was considered transient by Facebook.

This change vendors in the latest `go-glitch` to make use of the ability to add additional fields to the error. Here, we add a new `status.http` field holding the HTTP status code. This requires no changes to existing code to keep using `go-client`, but it allows one to probe for the field and take action based on the value, if present.