Go modules, TLS, and Clean-up
Update CircleCI configuration to use our common Go docker image. Update to use Go modules for the various commands.

Update the library to use Go modules for vendoring things into it.

Update `NewBaseClient` to not take in a `Roundtripper`. Instead, it takes in a `*tls.Config` for the client configuration. We default to the `http.DefaultTransport` for the `Roundtripper` used. If the `useTLS` value is `true` and a non-nil `tls.Config` is passed, we used that to build out a `http.Transport` for our `Roundtripper`.

Clean up some missing functionality around the `WithName` versions of functions.